requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
openai==1.3.7
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2